requirements.txt
# Basic data handling and plotting
pandas
matplotlib
seaborn

# System and audio metrics
psutil
pynvml
librosa

# Core ML/ASR Libraries
torch
torchaudio
transformers
openai-whisper

# For NVIDIA Models (Canary and Parakeet)
nemo_toolkit[asr]

# VAD Pipeline Dependencies
pathlib2
tqdm
datasets
soundfile
pydub

# Evaluation and metrics
jiwer 